What's new: 1) A quite a few of tweaks everywhere 2) Added  EnemyCombatAI 3) Redesigned EnemyMovementAI
It looks a bit messy, but I will sort things out when there's a bigger picture and at least Enemy system is working,
Enemies (Skeleton)  now properly roams between points, aim for the player when he's in reach and gets back to normal route when player is gone.
Attack system will be next for the enemy.